sink/mq(ticdc): split json.go and add callback
