+++ cdc/sink/mq/codec/canal_test.go
@@ -86,7 +86,7 @@ func TestCanalEventBatchEncoder(t *testing.T) {
 	for _, cs := range s.rowCases {
 		encoder := NewCanalEventBatchEncoder()
 		for _, row := range cs {
-			err := encoder.AppendRowChangedEvent(context.Background(), "", row)
+			err := encoder.AppendRowChangedEvent(context.Background(), "", row, nil)
 			require.Nil(t, err)
 		}
 		res := encoder.Build()
