+++ cdc/sink/mq/codec/codec_test.go
@@ -220,9 +220,9 @@ func TestJsonVsCraftVsPB(t *testing.T) {
 		craftEncoder.(*CraftEventBatchEncoder).maxBatchSize = 64
 		craftMessages := encodeRowCase(t, craftEncoder, cs)
 
-		jsonEncoder := NewJSONEventBatchEncoder()
-		jsonEncoder.(*JSONEventBatchEncoder).maxMessageBytes = 8192
-		jsonEncoder.(*JSONEventBatchEncoder).maxBatchSize = 64
+		jsonEncoder := newJSONEventBatchEncoder()
+		jsonEncoder.(*OpenProtocolBatchEncoder).maxMessageBytes = 8192
+		jsonEncoder.(*OpenProtocolBatchEncoder).maxBatchSize = 64
 		jsonMessages := encodeRowCase(t, jsonEncoder, cs)
 
 		protobuf1Messages := codecEncodeRowChangedPB1ToMessage(cs)
@@ -352,7 +352,7 @@ func codecEncodeRowChangedPB2(events []*model.RowChangedEvent) []byte {
 
 func codecEncodeRowCase(encoder EventBatchEncoder, events []*model.RowChangedEvent) ([]*MQMessage, error) {
 	for _, event := range events {
-		err := encoder.AppendRowChangedEvent(context.Background(), "", event)
+		err := encoder.AppendRowChangedEvent(context.Background(), "", event, nil)
 		if err != nil {
 			return nil, err
 		}
@@ -373,9 +373,9 @@ func init() {
 		panic(err)
 	}
 
-	encoder = NewJSONEventBatchEncoder()
-	encoder.(*JSONEventBatchEncoder).maxMessageBytes = 8192
-	encoder.(*JSONEventBatchEncoder).maxBatchSize = 64
+	encoder = newJSONEventBatchEncoder()
+	encoder.(*OpenProtocolBatchEncoder).maxMessageBytes = 8192
+	encoder.(*OpenProtocolBatchEncoder).maxBatchSize = 64
 	if codecJSONEncodedRowChanges, err = codecEncodeRowCase(encoder, codecBenchmarkRowChanges); err != nil {
 		panic(err)
 	}
@@ -394,9 +394,9 @@ func BenchmarkCraftEncoding(b *testing.B) {
 }
 
 func BenchmarkJsonEncoding(b *testing.B) {
-	encoder := NewJSONEventBatchEncoder()
-	encoder.(*JSONEventBatchEncoder).maxMessageBytes = 8192
-	encoder.(*JSONEventBatchEncoder).maxBatchSize = 64
+	encoder := newJSONEventBatchEncoder()
+	encoder.(*OpenProtocolBatchEncoder).maxMessageBytes = 8192
+	encoder.(*OpenProtocolBatchEncoder).maxBatchSize = 64
 	for i := 0; i < b.N; i++ {
 		_, _ = codecEncodeRowCase(encoder, codecBenchmarkRowChanges)
 	}
