+++ cdc/sink/mq/codec/maxwell_test.go
@@ -33,7 +33,7 @@ func TestMaxwellEventBatchCodec(t *testing.T) {
 	for _, cs := range rowCases {
 		encoder := newEncoder()
 		for _, row := range cs {
-			err := encoder.AppendRowChangedEvent(context.Background(), "", row)
+			err := encoder.AppendRowChangedEvent(context.Background(), "", row, nil)
 			require.Nil(t, err)
 		}
 		messages := encoder.Build()
