+++ cdc/sink/mq/codec/message.go
@@ -22,6 +22,13 @@ import (
 	"github.com/tikv/client-go/v2/oracle"
 )
 
+// MaxRecordOverhead is used to calculate message size by sarama kafka client.
+// reference: https://github.com/Shopify/sarama/blob/
+// 66521126c71c522c15a36663ae9cddc2b024c799/async_producer.go#L233
+// For TiCDC, minimum supported kafka version is `0.11.0.2`,
+// which will be treated as `version = 2` by sarama producer.
+const MaxRecordOverhead = 5*binary.MaxVarintLen32 + binary.MaxVarintLen64 + 1
+
 // MQMessage represents an MQ message to the mqSink
 type MQMessage struct {
 	Key       []byte
@@ -32,18 +39,14 @@ type MQMessage struct {
 	Type      model.MessageType // type
 	Protocol  config.Protocol   // protocol
 	rowsCount int               // rows in one MQ Message
+	Callback  func()            // Callback function will be called when the message is sent to the mqSink.
 }
-
-// maximumRecordOverhead is used to calculate ProducerMessage's byteSize by sarama kafka client.
-// reference: https://github.com/Shopify/sarama/blob/66521126c71c522c15a36663ae9cddc2b024c799/async_producer.go#L233
-// for TiCDC, minimum supported kafka version is `0.11.0.2`, which will be treated as `version = 2` by sarama producer.
-const maximumRecordOverhead = 5*binary.MaxVarintLen32 + binary.MaxVarintLen64 + 1
 
 // Length returns the expected size of the Kafka message
 // We didn't append any `Headers` when send the message, so ignore the calculations related to it.
 // If `ProducerMessage` Headers fields used, this method should also adjust.
 func (m *MQMessage) Length() int {
-	return len(m.Key) + len(m.Value) + maximumRecordOverhead
+	return len(m.Key) + len(m.Value) + MaxRecordOverhead
 }
 
 // PhysicalTime returns physical time part of Ts in time.Time
@@ -66,8 +69,8 @@ func (m *MQMessage) IncRowsCount() {
 	m.rowsCount++
 }
 
-func newDDLMQMessage(proto config.Protocol, key, value []byte, event *model.DDLEvent) *MQMessage {
-	return NewMQMessage(
+func newDDLMsg(proto config.Protocol, key, value []byte, event *model.DDLEvent) *MQMessage {
+	return newMsg(
 		proto,
 		key,
 		value,
@@ -78,13 +81,13 @@ func newDDLMQMessage(proto config.Protocol, key, value []byte, event *model.DDLE
 	)
 }
 
-func newResolvedMQMessage(proto config.Protocol, key, value []byte, ts uint64) *MQMessage {
-	return NewMQMessage(proto, key, value, ts, model.MessageTypeResolved, nil, nil)
+func newResolvedMsg(proto config.Protocol, key, value []byte, ts uint64) *MQMessage {
+	return newMsg(proto, key, value, ts, model.MessageTypeResolved, nil, nil)
 }
 
-// NewMQMessage should be used when creating a MQMessage struct.
+// newMsg should be used when creating a MQMessage struct.
 // It copies the input byte slices to avoid any surprises in asynchronous MQ writes.
-func NewMQMessage(
+func newMsg(
 	proto config.Protocol,
 	key []byte,
 	value []byte,
