+++ cdc/sink/mq/codec/message_test.go
@@ -46,7 +46,7 @@ func TestCreate(t *testing.T) {
 		CommitTs: 5678,
 	}
 
-	msg := NewMQMessage(config.ProtocolOpen, []byte("key1"), []byte("value1"), rowEvent.CommitTs, model.MessageTypeRow, &rowEvent.Table.Schema, &rowEvent.Table.Table)
+	msg := newMsg(config.ProtocolOpen, []byte("key1"), []byte("value1"), rowEvent.CommitTs, model.MessageTypeRow, &rowEvent.Table.Schema, &rowEvent.Table.Table)
 
 	require.Equal(t, []byte("key1"), msg.Key)
 	require.Equal(t, []byte("value1"), msg.Value)
@@ -94,7 +94,7 @@ func TestCreate(t *testing.T) {
 	ddlEvent := &model.DDLEvent{}
 	ddlEvent.FromJob(job, preTableInfo)
 
-	msg = newDDLMQMessage(config.ProtocolMaxwell, nil, []byte("value1"), ddlEvent)
+	msg = newDDLMsg(config.ProtocolMaxwell, nil, []byte("value1"), ddlEvent)
 	require.Nil(t, msg.Key)
 	require.Equal(t, []byte("value1"), msg.Value)
 	require.Equal(t, ddlEvent.CommitTs, msg.Ts)
@@ -103,7 +103,7 @@ func TestCreate(t *testing.T) {
 	require.Equal(t, ddlEvent.TableInfo.Table, *msg.Table)
 	require.Equal(t, config.ProtocolMaxwell, msg.Protocol)
 
-	msg = newResolvedMQMessage(config.ProtocolCanal, []byte("key1"), nil, 1234)
+	msg = newResolvedMsg(config.ProtocolCanal, []byte("key1"), nil, 1234)
 	require.Equal(t, []byte("key1"), msg.Key)
 	require.Nil(t, msg.Value)
 	require.Equal(t, uint64(1234), msg.Ts)
